Fix compilation errors
