+++ prov/verbs/src/ep_msg/verbs_cq_ep_msg.c
@@ -30,6 +30,46 @@
  * SOFTWARE.
  */
 
+#include <stdio.h>
+#include <stdlib.h>
+#include <sys/poll.h>
+
+#include <netinet/in.h>
+#include <infiniband/ib.h>
+
+#include <fi_enosys.h>
+#include <rdma/fi_eq.h>
+#include <prov/verbs/src/fi_verbs.h>
+
+extern struct fi_provider fi_ibv_prov;
+
+static ssize_t fi_ibv_cq_read_context(struct fid_cq *cq, void *buf, size_t count)
+{
+	struct fi_ibv_cq *_cq;
+	struct fi_cq_entry *entry = buf;
+	ssize_t ret = 0, i;
+
+	_cq = container_of(cq, struct fi_ibv_cq, cq_fid);
+	if (_cq->wc.status)
+		return -FI_EAVAIL;
+
+	for (i = 0; i < count; i++) {
+		ret = ibv_poll_cq(_cq->cq, 1, &_cq->wc);
+		if (ret <= 0)
+			break;
+
+		if (_cq->wc.status) {
+			ret = -FI_EAVAIL;
+			break;
+		}
+
+		entry->op_context = (void *) (uintptr_t) _cq->wc.wr_id;
+		entry += 1;
+	}
+
+	return i ? i : (ret ? ret : -FI_EAGAIN);
+}
+
 static ssize_t
 fi_ibv_cq_readerr(struct fid_cq *cq, struct fi_cq_err_entry *entry,
 		  uint64_t flags)
@@ -142,6 +182,42 @@ fi_ibv_cq_sread(struct fid_cq *cq, void *buf, size_t count, const void *cond,
 	return cur ? cur : ret;
 }
 
+static int
+fi_ibv_cq_signal(struct fid_cq *cq)
+{
+	struct fi_ibv_cq *_cq;
+	char data = '0';
+
+	_cq = container_of(cq, struct fi_ibv_cq, cq_fid);
+
+	if (write(_cq->signal_fd[1], &data, 1) != 1) {
+		FI_WARN(&fi_ibv_prov, FI_LOG_CQ, "Error signalling CQ\n");
+		return -errno;
+	}
+
+	return 0;
+}
+
+static const char *
+fi_ibv_cq_strerror(struct fid_cq *eq, int prov_errno, const void *err_data,
+		   char *buf, size_t len)
+{
+	if (buf && len)
+		strncpy(buf, ibv_wc_status_str(prov_errno), len);
+	return ibv_wc_status_str(prov_errno);
+}
+
+static struct fi_ops_cq fi_ibv_cq_context_ops = {
+	.size = sizeof(struct fi_ops_cq),
+	.read = fi_ibv_cq_read_context,
+	.readfrom = fi_no_cq_readfrom,
+	.readerr = fi_ibv_cq_readerr,
+	.sread = fi_ibv_cq_sread,
+	.sreadfrom = fi_no_cq_sreadfrom,
+	.signal = fi_ibv_cq_signal,
+	.strerror = fi_ibv_cq_strerror
+};
+
 static uint64_t fi_ibv_comp_flags(struct ibv_wc *wc)
 {
 	uint64_t flags = 0;
@@ -177,33 +253,6 @@ static uint64_t fi_ibv_comp_flags(struct ibv_wc *wc)
 	return flags;
 }
 
-static ssize_t fi_ibv_cq_read_context(struct fid_cq *cq, void *buf, size_t count)
-{
-	struct fi_ibv_cq *_cq;
-	struct fi_cq_entry *entry = buf;
-	ssize_t ret = 0, i;
-
-	_cq = container_of(cq, struct fi_ibv_cq, cq_fid);
-	if (_cq->wc.status)
-		return -FI_EAVAIL;
-
-	for (i = 0; i < count; i++) {
-		ret = ibv_poll_cq(_cq->cq, 1, &_cq->wc);
-		if (ret <= 0)
-			break;
-
-		if (_cq->wc.status) {
-			ret = -FI_EAVAIL;
-			break;
-		}
-
-		entry->op_context = (void *) (uintptr_t) _cq->wc.wr_id;
-		entry += 1;
-	}
-
-	return i ? i : (ret ? ret : -FI_EAGAIN);
-}
-
 static ssize_t fi_ibv_cq_read_msg(struct fid_cq *cq, void *buf, size_t count)
 {
 	struct fi_ibv_cq *_cq;
@@ -232,6 +281,17 @@ static ssize_t fi_ibv_cq_read_msg(struct fid_cq *cq, void *buf, size_t count)
 
 	return i ? i : (ret ? ret : -FI_EAGAIN);
 }
+
+static struct fi_ops_cq fi_ibv_cq_msg_ops = {
+	.size = sizeof(struct fi_ops_cq),
+	.read = fi_ibv_cq_read_msg,
+	.readfrom = fi_no_cq_readfrom,
+	.readerr = fi_ibv_cq_readerr,
+	.sread = fi_ibv_cq_sread,
+	.sreadfrom = fi_no_cq_sreadfrom,
+	.signal = fi_ibv_cq_signal,
+	.strerror = fi_ibv_cq_strerror
+};
 
 static ssize_t fi_ibv_cq_read_data(struct fid_cq *cq, void *buf, size_t count)
 {
@@ -271,53 +331,6 @@ static ssize_t fi_ibv_cq_read_data(struct fid_cq *cq, void *buf, size_t count)
 	return i ? i : (ret ? ret : -FI_EAGAIN);
 }
 
-static const char *
-fi_ibv_cq_strerror(struct fid_cq *eq, int prov_errno, const void *err_data,
-		   char *buf, size_t len)
-{
-	if (buf && len)
-		strncpy(buf, ibv_wc_status_str(prov_errno), len);
-	return ibv_wc_status_str(prov_errno);
-}
-
-static int
-fi_ibv_cq_signal(struct fid_cq *cq)
-{
-	struct fi_ibv_cq *_cq;
-	char data = '0';
-
-	_cq = container_of(cq, struct fi_ibv_cq, cq_fid);
-
-	if (write(_cq->signal_fd[1], &data, 1) != 1) {
-		FI_WARN(&fi_ibv_prov, FI_LOG_CQ, "Error signalling CQ\n");
-		return -errno;
-	}
-
-	return 0;
-}
-
-static struct fi_ops_cq fi_ibv_cq_context_ops = {
-	.size = sizeof(struct fi_ops_cq),
-	.read = fi_ibv_cq_read_context,
-	.readfrom = fi_no_cq_readfrom,
-	.readerr = fi_ibv_cq_readerr,
-	.sread = fi_ibv_cq_sread,
-	.sreadfrom = fi_no_cq_sreadfrom,
-	.signal = fi_ibv_cq_signal,
-	.strerror = fi_ibv_cq_strerror
-};
-
-static struct fi_ops_cq fi_ibv_cq_msg_ops = {
-	.size = sizeof(struct fi_ops_cq),
-	.read = fi_ibv_cq_read_msg,
-	.readfrom = fi_no_cq_readfrom,
-	.readerr = fi_ibv_cq_readerr,
-	.sread = fi_ibv_cq_sread,
-	.sreadfrom = fi_no_cq_sreadfrom,
-	.signal = fi_ibv_cq_signal,
-	.strerror = fi_ibv_cq_strerror
-};
-
 static struct fi_ops_cq fi_ibv_cq_data_ops = {
 	.size = sizeof(struct fi_ops_cq),
 	.read = fi_ibv_cq_read_data,
@@ -328,4 +341,158 @@ static struct fi_ops_cq fi_ibv_cq_data_ops = {
 	.signal = fi_ibv_cq_signal,
 	.strerror = fi_ibv_cq_strerror
 };
+
+static int fi_ibv_cq_control(fid_t fid, int command, void *arg)
+{
+	struct fi_ibv_cq *cq;
+	int ret = 0;
+
+	cq = container_of(fid, struct fi_ibv_cq, cq_fid.fid);
+	switch(command) {
+	case FI_GETWAIT:
+		if (!cq->channel) {
+			ret = -FI_ENODATA;
+			break;
+		}
+		*(int *) arg = cq->channel->fd;
+		break;
+	default:
+		ret = -FI_ENOSYS;
+		break;
+	}
+
+	return ret;
+}
+
+static int fi_ibv_cq_close(fid_t fid)
+{
+	struct fi_ibv_cq *cq;
+	int ret;
+
+	cq = container_of(fid, struct fi_ibv_cq, cq_fid.fid);
+	if (cq->cq) {
+		ret = ibv_destroy_cq(cq->cq);
+		if (ret)
+			return -ret;
+	}
+
+	if (cq->signal_fd[0]) {
+		close(cq->signal_fd[0]);
+	}
+	if (cq->signal_fd[1]) {
+		close(cq->signal_fd[1]);
+	}
+
+	if (cq->channel)
+		ibv_destroy_comp_channel(cq->channel);
+
+	free(cq);
+	return 0;
+}
+
+static struct fi_ops fi_ibv_cq_fi_ops = {
+	.size = sizeof(struct fi_ops),
+	.close = fi_ibv_cq_close,
+	.bind = fi_no_bind,
+	.control = fi_ibv_cq_control,
+	.ops_open = fi_no_ops_open,
+};
+
+int
+fi_ibv_cq_open(struct fid_domain *domain, struct fi_cq_attr *attr,
+	    struct fid_cq **cq, void *context)
+{
+	struct fi_ibv_cq *_cq;
+	int ret;
+
+	_cq = calloc(1, sizeof *_cq);
+	if (!_cq)
+		return -FI_ENOMEM;
+
+	_cq->domain = container_of(domain, struct fi_ibv_domain, domain_fid);
+
+	switch (attr->wait_obj) {
+	case FI_WAIT_UNSPEC:
+	case FI_WAIT_FD:
+		_cq->channel = ibv_create_comp_channel(_cq->domain->verbs);
+		if (!_cq->channel) {
+			ret = -errno;
+			goto err1;
+		}
+
+		ret = fi_fd_nonblock(_cq->channel->fd);
+		if (ret)
+			goto err2;
+
+		if (socketpair(AF_UNIX, SOCK_STREAM, 0, _cq->signal_fd)) {
+			ret = -errno;
+			goto err2;
+		}
+
+		ret = fi_fd_nonblock(_cq->signal_fd[0]);
+		if (ret)
+			goto err3;
+
+		break;
+	case FI_WAIT_NONE:
+		break;
+	default:
+		ret = -FI_ENOSYS;
+		goto err3;
+	}
+
+	_cq->cq = ibv_create_cq(_cq->domain->verbs, attr->size, _cq,
+				_cq->channel, attr->signaling_vector);
+	if (!_cq->cq) {
+		ret = -errno;
+		goto err3;
+	}
+
+	if (_cq->channel) {
+		ret = ibv_req_notify_cq(_cq->cq, 0);
+		if (ret) {
+			FI_WARN(&fi_ibv_prov, FI_LOG_CQ, "ibv_req_notify_cq failed\n");
+			goto err4;
+		}
+	}
+
+	_cq->flags |= attr->flags;
+	_cq->wait_cond = attr->wait_cond;
+	_cq->cq_fid.fid.fclass = FI_CLASS_CQ;
+	_cq->cq_fid.fid.context = context;
+	_cq->cq_fid.fid.ops = &fi_ibv_cq_fi_ops;
+
+	switch (attr->format) {
+	case FI_CQ_FORMAT_CONTEXT:
+		_cq->cq_fid.ops = &fi_ibv_cq_context_ops;
+		_cq->entry_size = sizeof(struct fi_cq_entry);
+		break;
+	case FI_CQ_FORMAT_MSG:
+		_cq->cq_fid.ops = &fi_ibv_cq_msg_ops;
+		_cq->entry_size = sizeof(struct fi_cq_msg_entry);
+		break;
+	case FI_CQ_FORMAT_DATA:
+		_cq->cq_fid.ops = &fi_ibv_cq_data_ops;
+		_cq->entry_size = sizeof(struct fi_cq_data_entry);
+		break;
+	default:
+		ret = -FI_ENOSYS;
+		goto err4;
+	}
+
+	*cq = &_cq->cq_fid;
+	return 0;
+
+err4:
+	ibv_destroy_cq(_cq->cq);
+err3:
+	close(_cq->signal_fd[0]);
+	close(_cq->signal_fd[1]);
+err2:
+	if (_cq->channel)
+		ibv_destroy_comp_channel(_cq->channel);
+err1:
+	free(_cq);
+	return ret;
+}
