+++ prov/verbs/src/verbs_info.c
@@ -30,10 +30,25 @@
  * SOFTWARE.
  */
 
-static struct fi_info *verbs_info = NULL;
+#include <stddef.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <pthread.h>
+
+#include <rdma/rdma_cma.h>
+#include <infiniband/ib.h>
+
+#include <rdma/fabric.h>
+#include <prov/verbs/src/fi_verbs.h>
+#include <prov/verbs/src/verbs_ep.h>
+#include <prov/verbs/src/verbs_checks.h>
+
+extern struct fi_provider fi_ibv_prov;
+
+struct fi_info *verbs_info = NULL;
+
 static pthread_mutex_t verbs_info_lock = PTHREAD_MUTEX_INITIALIZER;
 
-static const char *local_node = "localhost";
 static char def_tx_ctx_size[16] = "384";
 static char def_rx_ctx_size[16] = "384";
 static char def_tx_iov_limit[16] = "4";
@@ -129,64 +144,6 @@ static int fi_ibv_check_hints(const struct fi_info *hints,
 		ret = fi_ibv_check_tx_attr(hints->tx_attr, hints, info);
 		if (ret)
 			return ret;
-	}
-
-	return 0;
-}
-
-static int fi_ibv_fi_to_rai(const struct fi_info *fi, uint64_t flags, struct rdma_addrinfo *rai)
-{
-	memset(rai, 0, sizeof *rai);
-	if (flags & FI_SOURCE)
-		rai->ai_flags = RAI_PASSIVE;
-	if (flags & FI_NUMERICHOST)
-		rai->ai_flags |= RAI_NUMERICHOST;
-
-	rai->ai_qp_type = IBV_QPT_RC;
-	rai->ai_port_space = RDMA_PS_TCP;
-
-	if (!fi)
-		return 0;
-
-	switch(fi->addr_format) {
-	case FI_SOCKADDR_IN:
-		rai->ai_family = AF_INET;
-		rai->ai_flags |= RAI_FAMILY;
-		break;
-	case FI_SOCKADDR_IN6:
-		rai->ai_family = AF_INET6;
-		rai->ai_flags |= RAI_FAMILY;
-		break;
-	case FI_SOCKADDR_IB:
-		rai->ai_family = AF_IB;
-		rai->ai_flags |= RAI_FAMILY;
-		break;
-	case FI_SOCKADDR:
-		if (fi->src_addrlen) {
-			rai->ai_family = ((struct sockaddr *)fi->src_addr)->sa_family;
-			rai->ai_flags |= RAI_FAMILY;
-		} else if (fi->dest_addrlen) {
-			rai->ai_family = ((struct sockaddr *)fi->dest_addr)->sa_family;
-			rai->ai_flags |= RAI_FAMILY;
-		}
-		break;
-	case FI_FORMAT_UNSPEC:
-		break;
-	default:
-		VERBS_INFO(FI_LOG_FABRIC, "Unknown fi->addr_format\n");
-	}
-
-	if (fi->src_addrlen) {
-		if (!(rai->ai_src_addr = malloc(fi->src_addrlen)))
-			return -FI_ENOMEM;
-		memcpy(rai->ai_src_addr, fi->src_addr, fi->src_addrlen);
-		rai->ai_src_len = fi->src_addrlen;
-	}
-	if (fi->dest_addrlen) {
-		if (!(rai->ai_dst_addr = malloc(fi->dest_addrlen)))
-			return -FI_ENOMEM;
-		memcpy(rai->ai_dst_addr, fi->dest_addr, fi->dest_addrlen);
-		rai->ai_dst_len = fi->dest_addrlen;
 	}
 
 	return 0;
@@ -429,7 +386,7 @@ err:
 	return ret;
 }
 
-static int fi_ibv_init_info(void)
+int fi_ibv_init_info(void)
 {
 	struct ibv_context **ctx_list;
 	struct fi_info *fi = NULL, *tail = NULL;
@@ -474,74 +431,31 @@ unlock:
 	return ret;
 }
 
-static int
-fi_ibv_create_ep(const char *node, const char *service,
-		 uint64_t flags, const struct fi_info *hints,
-		 struct rdma_addrinfo **rai, struct rdma_cm_id **id)
+void fi_ibv_update_info(const struct fi_info *hints, struct fi_info *info)
 {
-	struct rdma_addrinfo rai_hints, *_rai;
-	struct rdma_addrinfo **rai_current;
-	int ret;
-
-	ret = fi_ibv_fi_to_rai(hints, flags, &rai_hints);
-	if (ret)
-		goto out;
-
-	if (!node && !rai_hints.ai_dst_addr) {
-		if (!rai_hints.ai_src_addr && !service) {
-			node = local_node;
+	if (hints) {
+		if (hints->ep_attr) {
+			if (hints->ep_attr->tx_ctx_cnt)
+				info->ep_attr->tx_ctx_cnt = hints->ep_attr->tx_ctx_cnt;
+			if (hints->ep_attr->rx_ctx_cnt)
+				info->ep_attr->rx_ctx_cnt = hints->ep_attr->rx_ctx_cnt;
 		}
-		rai_hints.ai_flags |= RAI_PASSIVE;
-	}
 
-	ret = rdma_getaddrinfo((char *) node, (char *) service,
-				&rai_hints, &_rai);
-	if (ret) {
-		VERBS_INFO_ERRNO(FI_LOG_FABRIC, "rdma_getaddrinfo", errno);
-		ret = -errno;
-		goto out;
-	}
+		if (hints->tx_attr)
+			info->tx_attr->op_flags = hints->tx_attr->op_flags;
 
-	/*
-	 * If caller requested rai, remove ib_rai entries added by IBACM to
-	 * prevent wrong ib_connect_hdr from being sent in connect request.
-	 */
-	if (rai && hints && (hints->addr_format != FI_SOCKADDR_IB)) {
-		for (rai_current = &_rai; *rai_current;) {
-			struct rdma_addrinfo *rai_next;
-			if ((*rai_current)->ai_family == AF_IB) {
-				rai_next = (*rai_current)->ai_next;
-				(*rai_current)->ai_next = NULL;
-				rdma_freeaddrinfo(*rai_current);
-				*rai_current = rai_next;
-				continue;
-			}
-			rai_current = &(*rai_current)->ai_next;
-		}
-	}
+		if (hints->rx_attr)
+			info->rx_attr->op_flags = hints->rx_attr->op_flags;
 
-	ret = rdma_create_ep(id, _rai, NULL, NULL);
-	if (ret) {
-		VERBS_INFO_ERRNO(FI_LOG_FABRIC, "rdma_create_ep", errno);
-		ret = -errno;
-		goto err;
+		if (hints->handle)
+			info->handle = hints->handle;
+	} else {
+		info->tx_attr->op_flags = 0;
+		info->rx_attr->op_flags = 0;
 	}
-
-	if (rai) {
-		*rai = _rai;
-		goto out;
-	}
-err:
-	rdma_freeaddrinfo(_rai);
-out:
-	if (rai_hints.ai_src_addr)
-		free(rai_hints.ai_src_addr);
-	if (rai_hints.ai_dst_addr)
-		free(rai_hints.ai_dst_addr);
-	return ret;
 }
 
-static struct fi_info *fi_ibv_search_verbs_info(const char *fabric_name,
+struct fi_info *fi_ibv_search_verbs_info(const char *fabric_name,
 		const char *domain_name)
 {
 	struct fi_info *info;
@@ -602,7 +516,7 @@ err1:
 	return ret;
 }
 
-static int fi_ibv_getinfo(uint32_t version, const char *node, const char *service,
+int fi_ibv_getinfo(uint32_t version, const char *node, const char *service,
 			  uint64_t flags, struct fi_info *hints, struct fi_info **info)
 {
 	struct rdma_cm_id *id;
