+++ prov/verbs/src/verbs_utils.c
@@ -30,7 +30,38 @@
  * SOFTWARE.
  */
 
-static int fi_ibv_copy_addr(void *dst_addr, size_t *dst_addrlen, void *src_addr)
+
+#include <prov/verbs/src/verbs_utils.h>
+
+#include <alloca.h>
+#include <sys/socket.h>
+#include <arpa/inet.h>
+#include <netinet/in.h>
+#include <infiniband/ib.h>
+#include <rdma/rdma_cma.h>
+
+#include <rdma/fi_errno.h>
+
+#include <prov/verbs/src/fi_verbs.h>
+
+int fi_ibv_sockaddr_len(struct sockaddr *addr)
+{
+	if (!addr)
+		return 0;
+
+	switch (addr->sa_family) {
+	case AF_INET:
+		return sizeof(struct sockaddr_in);
+	case AF_INET6:
+		return sizeof(struct sockaddr_in6);
+	case AF_IB:
+		return sizeof(struct sockaddr_ib);
+	default:
+		return 0;
+	}
+}
+
+int fi_ibv_copy_addr(void *dst_addr, size_t *dst_addrlen, void *src_addr)
 {
 	size_t src_addrlen = fi_ibv_sockaddr_len(src_addr);
 
@@ -48,3 +79,51 @@ static int fi_ibv_copy_addr(void *dst_addr, size_t *dst_addrlen, void *src_addr)
 	return 0;
 }
 
+ssize_t fi_ibv_send(struct fi_ibv_msg_ep *ep, struct ibv_send_wr *wr, size_t len,
+		int count, void *context)
+{
+	struct ibv_send_wr *bad_wr;
+	int ret;
+
+	wr->num_sge = count;
+	wr->wr_id = (uintptr_t) context;
+
+	ret = ibv_post_send(ep->id->qp, wr, &bad_wr);
+	switch (ret) {
+	case ENOMEM:
+		return -FI_EAGAIN;
+	case -1:
+		/* Deal with non-compliant libibverbs drivers which set errno
+		 * instead of directly returning the error value */
+		return (errno == ENOMEM) ? -FI_EAGAIN : -errno;
+	default:
+		return -ret;
+	}
+}
+
+ssize_t fi_ibv_send_buf(struct fi_ibv_msg_ep *ep, struct ibv_send_wr *wr,
+		const void *buf, size_t len, void *desc, void *context)
+{
+	struct ibv_sge sge;
+
+	fi_ibv_set_sge(sge, buf, len, desc);
+	wr->sg_list = &sge;
+
+	return fi_ibv_send(ep, wr, len, 1, context);
+}
+
+ssize_t fi_ibv_send_iov_flags(struct fi_ibv_msg_ep *ep, struct ibv_send_wr *wr,
+		const struct iovec *iov, void **desc, int count, void *context,
+		uint64_t flags)
+{
+	size_t len = 0;
+
+	if (!desc)
+		fi_ibv_set_sge_iov_inline(wr->sg_list, iov, count, len);
+	else
+		fi_ibv_set_sge_iov(wr->sg_list, iov, count, desc, len);
+
+	wr->send_flags = VERBS_INJECT_FLAGS(ep, len, flags) | VERBS_COMP_FLAGS(ep, flags);
+
+	return fi_ibv_send(ep, wr, len, count, context);
+}
